+++ src/api/qualicharge/factories/dynamic.py
@@ -1,8 +1,9 @@
 """QualiCharge dynamic factories."""
 
-from datetime import timezone
+from datetime import datetime, timezone
 
 from polyfactory import Use
+from polyfactory.decorators import post_generated
 from polyfactory.factories.dataclass_factory import DataclassFactory
 from polyfactory.factories.pydantic_factory import ModelFactory
 
@@ -27,12 +28,18 @@ class SessionCreateFactory(ModelFactory[SessionCreate]):
             tzinfo=timezone.utc,
         )
     )
-    end = Use(
-        lambda: DataclassFactory.__faker__.date_time_between(
-            start_date=f"-{settings.API_MAX_SESSION_AGE - 3600}s",
+
+    @post_generated
+    @classmethod
+    def end(cls, start: datetime):
+        """Ensure session duration is less than a week."""
+        from_ = int((datetime.now(timezone.utc) - start).total_seconds()) + 1
+        to_ = from_ - (60 * 60 * 24 * 7)  # a week
+        return DataclassFactory.__faker__.date_time_between(
+            start_date=f"-{from_}s",
+            end_date=f"-{to_}s",
             tzinfo=timezone.utc,
         )
-    )
 
 
 class StatusCreateFactory(ModelFactory[StatusCreate]):