+++ src/api/tests/models/test_static.py
@@ -1,12 +1,14 @@
 """QualiCharge static models tests."""
 
+import json
 from datetime import datetime, timedelta, timezone
 
 import pytest
+from pydantic import TypeAdapter, ValidationError
 from pydantic_extra_types.coordinate import Coordinate
 
 from qualicharge.factories.static import StatiqueFactory
-from qualicharge.models.static import Statique
+from qualicharge.models.static import RaccordementEnum, Siren, Statique
 
 
 def test_statique_model_coordonneesXY():
@@ -63,6 +65,148 @@ def test_statique_model_french_phone_numbers(phone_number):
     assert statique.telephone_operateur == "tel:+33-1-44-27-63-50"
 
 
+@pytest.mark.parametrize(
+    "value",
+    [
+        "",  # too short
+        "1",
+        "1234567890",  # too long
+    ],
+)
+def test_siren_check_invalid_length(value):
+    """Test the `Siren` type validation."""
+    siren = TypeAdapter(Siren)
+    with pytest.raises(ValidationError, match="1 validation error"):
+        siren.validate_python(value)
+
+
+@pytest.mark.parametrize(
+    "value",
+    [
+        "abcdefghi",  # not a number
+        "1234e6789",
+        "000000000",  # blacklist
+        "123456789",  # invalid
+    ],
+)
+def test_siren_check_invalid(value):
+    """Test the `Siren` type validation."""
+    siren = TypeAdapter(Siren)
+    with pytest.raises(ValidationError, match=f"{value} is not a valid SIREN number"):
+        siren.validate_python(value)
+
+
+@pytest.mark.parametrize(
+    "value",
+    [
+        "256300146",
+        "524335262",
+        "531680445",
+        "539188169",
+        "818943938",
+        "835124280",
+        "838436145",
+        "842718512",
+        "844192443",
+        "848778429",
+        "885354860",
+        "891118473",
+        "891624884",
+        "897849923",
+        "909073363",
+        "911482628",
+        "917546251",
+        "932449226",
+        "934419615",
+        "940861826",
+        "953900123",
+        "979758828",
+        "983504002",
+        "984586875",
+        "985330364",
+    ],
+)
+def test_siren_check_valid(value):
+    """Test the `Siren` custom type."""
+    siren = TypeAdapter(Siren)
+    assert siren.validate_python(value) == value
+
+
+@pytest.mark.parametrize(
+    "siren",
+    [None, 42],
+)
+def test_statique_amenageur_siren_invalid_type(siren):
+    """Test the statique model `siren_amenageur` field (type)."""
+    with pytest.raises(
+        ValueError,
+        match=("Input should be a valid string"),
+    ):
+        StatiqueFactory.build(siren_amenageur=siren)
+
+
+@pytest.mark.parametrize(
+    "siren",
+    ["", "abcdef", "0000"],
+)
+def test_statique_amenageur_siren_invalid_str(siren):
+    """Test the statique model `siren_amenageur` field (string length)."""
+    with pytest.raises(
+        ValueError,
+        match=("String should have at least 9 characters"),
+    ):
+        StatiqueFactory.build(siren_amenageur=siren)
+
+
+@pytest.mark.parametrize(
+    "siren",
+    ["000000000", "123456789", "abcdefghi"],
+)
+def test_statique_amenageur_siren_invalid(siren):
+    """Test the statique model `siren_amenageur` field (SIREN check)."""
+    with pytest.raises(
+        ValueError,
+        match=(f"{siren} is not a valid SIREN number"),
+    ):
+        StatiqueFactory.build(siren_amenageur=siren)
+
+
+@pytest.mark.parametrize(
+    "siren",
+    [
+        "256300146",
+        "524335262",
+        "531680445",
+        "539188169",
+        "818943938",
+        "835124280",
+        "838436145",
+        "842718512",
+        "844192443",
+        "848778429",
+        "885354860",
+        "891118473",
+        "891624884",
+        "897849923",
+        "909073363",
+        "911482628",
+        "917546251",
+        "932449226",
+        "934419615",
+        "940861826",
+        "953900123",
+        "979758828",
+        "983504002",
+        "984586875",
+        "985330364",
+    ],
+)
+def test_statique_amenageur_siren_valid(siren):
+    """Test the `Siren` custom type."""
+    statique = StatiqueFactory.build(siren_amenageur=siren)
+    assert statique.siren_amenageur == siren
+
+
 def test_statique_model_json_schema():
     """Test the Statique model generated JSON schema."""
     schema = Statique.model_json_schema()
@@ -76,7 +220,7 @@ def test_statique_model_json_schema():
     assert schema["properties"]["coordonneesXY"]["examples"] == ["[12.3, 41.5]"]
 
 
-def test_statique_model_afirev_previx_check():
+def test_statique_model_afirev_prefix_check():
     """Test the id_pdc/station_itinerance consistency."""
     with pytest.raises(
         ValueError,
@@ -108,6 +252,38 @@ def test_statique_model_num_pdl():
 
     with pytest.raises(ValueError, match="String should have at most 64 characters"):
         StatiqueFactory.build(num_pdl="a" * 65)
+
+    # num_pdl is required for direct connections
+    statique = StatiqueFactory.build(raccordement=RaccordementEnum.DIRECT)
+    with pytest.raises(
+        ValueError, match="A PDL number is required for direct connections."
+    ):
+        Statique(
+            num_pdl=None, **json.loads(statique.model_dump_json(exclude={"num_pdl"}))
+        )
+    with pytest.raises(
+        ValueError, match="A PDL number is required for direct connections."
+    ):
+        Statique(
+            num_pdl="", **json.loads(statique.model_dump_json(exclude={"num_pdl"}))
+        )
+
+
+def test_statique_model_puissance_nominale():
+    """Test statique model puissance_nominale values."""
+    statique = StatiqueFactory.build(puissance_nominale=200.0)
+    with pytest.raises(
+        ValueError, match="Input should be greater than or equal to 1.3"
+    ):
+        Statique(
+            puissance_nominale=1.0,
+            **json.loads(statique.model_dump_json(exclude={"puissance_nominale"})),
+        )
+    with pytest.raises(ValueError, match="Input should be less than or equal to 4000"):
+        Statique(
+            puissance_nominale=4001.0,
+            **json.loads(statique.model_dump_json(exclude={"puissance_nominale"})),
+        )
 
 
 def test_statique_model_date_maj():